backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
assemblyai==0.21.0
openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.27.1